tools/requirements.txt
paramiko
pyyaml
python-digitalocean
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-dotenv